- Implement first cut of Migration.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@2778 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/stored/stored.c
@@ -215,7 +215,7 @@ int main (int argc, char *argv[])
    /* Ensure that Volume Session Time and Id are both
     * set and are both non-zero.
     */
-   VolSessionTime = (long)daemon_start_time;
+   VolSessionTime = (uint32_t)daemon_start_time;
    if (VolSessionTime == 0) { /* paranoid */
       Jmsg0(NULL, M_ABORT, 0, _("Volume Session Time is ZERO!\n"));
    }
